feat: fix guards and make pagination

+++ src/article/article.resolver.ts
@@ -3,20 +3,34 @@ import { ArticleService } from './article.service';
 import { Article } from './entities/article.entity';
 import { CreateArticleInput } from './dto/create-article.input';
 import { UpdateArticleInput } from './dto/update-article.input';
-import { UseGuards } from '@nestjs/common';
-import { FacultyAdmin } from 'src/guards/facultyAdmin.guard';
-import { PsyhopeAdmin } from 'src/guards/psyhopeAdmin.guard';
+import { UseGuards, HttpException, HttpStatus } from '@nestjs/common';
 import { LoggedIn } from 'src/guards/loggedIn.guard';
+import { CurrentUser } from 'src/auth/decorator/currentUser.decorator';
+import { JwtPayload } from 'src/auth/interfaces/jwt.payload';
+import { UserRepositories } from 'src/models/user.repo';
+import { Role } from '@prisma/client';
 
 @Resolver(() => Article)
 export class ArticleResolver {
-  constructor(private readonly articleService: ArticleService) {}
+  constructor(
+    private readonly articleService: ArticleService,
+    private readonly userRepo: UserRepositories,
+  ) {}
 
   @Mutation(() => Article)
-  @UseGuards(LoggedIn, FacultyAdmin, PsyhopeAdmin)
-  createArticle(
+  @UseGuards(LoggedIn)
+  async createArticle(
     @Args('createArticleInput') createArticleInput: CreateArticleInput,
+    @CurrentUser() user: JwtPayload,
   ) {
+    const { account } = await this.userRepo.findById(user.sub);
+
+    if (
+      account.role != Role.PSYHOPE_ADMIN &&
+      account.role != Role.FACULTY_ADMIN
+    )
+      throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
+
     return this.articleService.create(createArticleInput);
   }
 
@@ -36,16 +50,46 @@ export class ArticleResolver {
   }
 
   @Mutation(() => Article)
-  @UseGuards(LoggedIn, FacultyAdmin, PsyhopeAdmin)
-  updateArticle(
+  @UseGuards(LoggedIn)
+  async updateArticle(
     @Args('updateArticleInput') updateArticleInput: UpdateArticleInput,
+    @CurrentUser() user: JwtPayload,
   ) {
+    const { account } = await this.userRepo.findById(user.sub);
+
+    if (
+      account.role != Role.PSYHOPE_ADMIN &&
+      account.role != Role.FACULTY_ADMIN
+    )
+      throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
+
     return this.articleService.update(updateArticleInput);
   }
 
   @Mutation(() => Article)
-  @UseGuards(LoggedIn, FacultyAdmin, PsyhopeAdmin)
-  removeArticle(@Args('id', { type: () => Int }) id: number) {
+  @UseGuards(LoggedIn)
+  async removeArticle(
+    @Args('id', { type: () => Int }) id: number,
+    @CurrentUser() user: JwtPayload,
+  ) {
+    const { account } = await this.userRepo.findById(user.sub);
+
+    if (
+      account.role != Role.PSYHOPE_ADMIN &&
+      account.role != Role.FACULTY_ADMIN
+    )
+      throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
+
     return this.articleService.remove(id);
+  }
+
+  @Query(() => Article)
+  findByLimitArticle(@Args('limit', { type: () => Int }) limit: number) {
+    return this.articleService.findByLimit(limit);
+  }
+
+  @Query(() => Number)
+  countArticle() {
+    return this.articleService.count();
   }
 }
